producer/requirements.txt
kafka-python==2.0.2
watchdog==0.8.2